ベースクラスEffectComponentの実装 [Feature] 基本的なエフェクトシステム Fixes #10
